pyproject.toml
[tool.poetry]
name = "course-project-python"
version = "0.1.0"
description = ""
authors = ["Agata <agagor1961@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
logging = "^0.4.9.6"
openpyxl = "^3.1.4"
pandas = "^2.2.2"
xlrd = "^2.0.1"
requests = "^2.32.3"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
pytest-cov = "^5.0.0"


[tool.poetry.group.lint.dependencies]
black = "^24.4.2"
isort = "^5.13.2"
flake8 = "^7.1.0"
mypy = "^1.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
exclude = 'venv'

[tool.black]
line-length = 119
exclude = '''
(
  /(
      \.eggs
    | \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | dist
  )/
  | foo.py
)
'''

[tool.isort]
line_length = 119